[bug fix] Move gql_platform.graphene to gql_platform.graphene_schema.
This avoids imports of graphene the third party package from
colliding with imports of graphene the project package.

Because the python path lists the source of the script before
site-packages, without this fix, import of graphene may end up
incorrectly importing package project.gql_platform.graphene.
