+++ graphene-django-backend/project/gql_platform/graphene_schema/examples.gql
